pyproject.toml
[project]
name = "ulsdk"
version = "0.1.0"
description = "A Python SDK for the UrbanLogiq API"
dependencies = [
  "flatbuffers",
  "pynacl",
  "pyarrow",
  "requests",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "ulsdk"
version = "0.1.0"
description = "A Python SDK for the UrbanLogiq API"
authors = ["UrbanLogiq <customersuccess@urbanlogiq.com>"]

[tool.poetry.dependencies]
python = ">=3.12, <4"
pynacl = "1.5.0"
requests = "^2.32.0"
pyarrow = "^17"
pyarrow-stubs = "^10.0.1.9"
flatbuffers = "^22.12.6"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.377"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[tool.pyright]
reportAttributeAccessIssue = false
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = true
reportUntypedClassDecorator = true
pythonVersion = "3.12"
executionEnvironments = [
  { root = "." },
]

[tool.setuptools.packages.find]
where = ["python"]
